Fix clippy-wrapper test race condition.
